Add `foo` and `bar` methods
We need `foo` and `bar` methods in order to console.log a string
representing which method is executed. This will help us for better
debugging. This methods are not used atm, but they will be in some near
future.

Things to consider:
- use them wisely
- use them for debugging purposes only
- you can delete them at any time

Contact me if you need more explaination about these methods.